# Replace PL/PGSQL executor with native SPI