--- a/Cargo.toml
+++ b/Cargo.toml
@@ -9,10 +9,11 @@ crate-type = ["cdylib"]
 [features]
 default = ["pg14"]
 pg14 = ["pgx/pg14", "pgx-tests/pg14" ]
+pg15 = ["pgx/pg15", "pgx-tests/pg15" ]
 pg_test = []
 
 [dependencies]
-pgx = "~0.5.6"
+pgx = "=0.6.0-alpha.0"
 graphql-parser = "0.4"
 serde = "1.0"
 serde_json = "1.0"
@@ -23,11 +24,12 @@ uuid = "1"
 regex = "1"
 base64 = "0.13"
 lazy_static = "1"
+pgx-contrib-spiext = { git = "https://github.com/supabase/pgx-contrib-spiext", rev = "28642a86da1b5b361c2724828ebf22cfad55e557" }
 
 
 
 [dev-dependencies]
-pgx-tests = "~0.5.6"
+pgx-tests = "=0.6.0-alpha.0"
 
 [profile.dev]
 panic = "unwind"