--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,7 +13,7 @@ pg15 = ["pgx/pg15", "pgx-tests/pg15" ]
 pg_test = []
 
 [dependencies]
-pgx = "=0.6.0-alpha.0"
+pgx = "=0.6.0-alpha.1"
 graphql-parser = "0.4"
 serde = "1.0"
 serde_json = "1.0"