--- a/sql/raise_exception.sql
+++ b/sql/raise_exception.sql
@@ -1,0 +1,8 @@
+create or replace function graphql.exception(message text)
+    returns text
+    language plpgsql
+as $$
+begin
+    raise exception using errcode='22000', message=message;
+end;
+$$;