--- a/src/transpile.rs
+++ b/src/transpile.rs
@@ -566,20 +566,30 @@ impl OrderByBuilder {
     }
 }
 
-pub fn json_to_literal(val: &serde_json::Value) -> String {
+pub fn json_to_text_datum(val: &serde_json::Value) -> Option<pg_sys::Datum> {
+    let null: Option<i32> = None;
     match val {
-        serde_json::Value::Null => "null".to_string(),
-        serde_json::Value::Bool(_) => quote_literal(&val.to_string()),
-        serde_json::Value::String(x) => quote_literal(&x.to_string()),
-        serde_json::Value::Number(x) => quote_literal(&x.to_string()),
+        serde_json::Value::Null => null.into_datum(),
+        serde_json::Value::Bool(_) => val.to_string().into_datum(),
+        serde_json::Value::String(x) => x.into_datum(),
+        serde_json::Value::Number(x) => x.to_string().into_datum(),
+        serde_json::Value::Array(xarr) => {
+            let inner_vals: Vec<Option<String>> = xarr
+                .iter()
+                .map(|x| match x {
+                    serde_json::Value::Null => None,
+                    serde_json::Value::Bool(x) => Some(x.to_string()),
+                    serde_json::Value::String(x) => Some(x.to_string()),
+                    serde_json::Value::Number(x) => Some(x.to_string()),
+                    serde_json::Value::Array(_) => panic!("Unexpected array in input value array"),
+                    serde_json::Value::Object(_) => {
+                        panic!("Unexpected object in input value array")
+                    }
+                })
+                .collect();
+            inner_vals.into_datum()
+        }
         // Should this ever happen? json input is escaped so it would be a string.
-        serde_json::Value::Array(xarr) => format!(
-            "array[{}]",
-            xarr.iter()
-                .map(json_to_literal)
-                .collect::<Vec<String>>()
-                .join(", ")
-        ),
         serde_json::Value::Object(_) => panic!("Unexpected object in input value"),
     }
 }
@@ -589,9 +599,17 @@ pub struct ParamContext {
 }
 
 impl ParamContext {
+    // Pushes a parameter into the context and returns a SQL clause to reference it
+    //fn clause_for(&mut self, param: (PgOid, Option<pg_sys::Datum>)) -> String {
     fn clause_for(&mut self, value: &serde_json::Value, type_name: &String) -> String {
-        let safe_literal = json_to_literal(value);
-        format!("({}::{})", safe_literal, type_name)
+        let type_oid = match type_name.ends_with("[]") {
+            true => PgOid::from(1009), // text[]
+            false => PgOid::from(25),  // text
+        };
+
+        let val_datum = json_to_text_datum(value);
+        self.params.push((type_oid, val_datum));
+        format!("(${}::{})", self.params.len(), type_name)
     }
 }