## What kind of change does this PR introduce?
- Updates pgx to 0.6.0-alpha.0, which adds support for
   - pg15
   - [pgx-contrib-spx](https://github.com/supabase/pgx-contrib-spiext)

Using pgx-contrib-spiext, we:
-  Replace the PL/PGSQL Sequential Executor hack to allow executing multiple mutations in a single transaction while returning error messages with the native SPI that has support for those capabilities.
- Replace the literal binds required by the PL/PGSQL Sequential executor hack with prepared statements that take arguments

## Performance Impact

- 47% increase in request throughput
- 5x reduction in 99% response time (25ms to 5ms)

Before
```
Concurrency Level:      20
Time taken for tests:   0.932 seconds
Complete requests:      5000
Failed requests:        0
Total transferred:      1080000 bytes
Total body sent:        1160000
HTML transferred:       325000 bytes
Requests per second:    5366.30 [#/sec] (mean)
Time per request:       3.727 [ms] (mean)
Time per request:       0.186 [ms] (mean, across all concurrent requests)
Transfer rate:          1131.95 [Kbytes/sec] received
                        1215.80 kb/s sent
                        2347.76 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:     1    4   3.9      3      69
Waiting:        1    4   3.9      3      69
Total:          1    4   3.9      3      69

Percentage of the requests served within a certain time (ms)
  50%      3
  66%      4
  75%      4
  80%      4
  90%      5
  95%      7
  98%     16
  99%     25
 100%     69 (longest request)
```

After
```
Concurrency Level:      20
Time taken for tests:   0.630 seconds
Complete requests:      5000
Failed requests:        0
Total transferred:      1210000 bytes
Total body sent:        1160000
HTML transferred:       325000 bytes
Requests per second:    7940.59 [#/sec] (mean)
Time per request:       2.519 [ms] (mean)
Time per request:       0.126 [ms] (mean, across all concurrent requests)
Transfer rate:          1876.59 [Kbytes/sec] received
                        1799.04 kb/s sent
                        3675.63 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       2
Processing:     1    2   0.8      2      13
Waiting:        1    2   0.8      2      13
Total:          1    2   0.9      2      13

Percentage of the requests served within a certain time (ms)
  50%      2
  66%      3
  75%      3
  80%      3
  90%      4
  95%      4
  98%      5
  99%      5
 100%     13 (longest request)
```